closed-env-package/requirements.txt
# Зависимости для семантического агента
# Установка: pip install -r requirements.txt
#
# ВАЖНО: В закрытом контуре без интернета скачайте .whl файлы заранее
# и установите: pip install --no-index --find-links=./wheels -r requirements.txt

# LLM
langchain-gigachat>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain>=0.2.0

# Форматирование таблиц
tabulate>=0.9.0
pandas>=2.0.0

# Векторная БД
faiss-cpu>=1.7.4

# Эмбеддинги
sentence-transformers>=2.2.0

# HTTP-клиент (для Cube API)
httpx>=0.25.0

# PostgreSQL
psycopg2-binary>=2.9.0

# Конфигурация
pyyaml>=6.0
python-dotenv>=1.0.0

# Jupyter (обычно уже установлен)
# jupyterlab>=4.0
